+++ src/main/java/com/dodream/job/application/JobService.java
@@ -1,12 +1,15 @@
 package com.dodream.job.application;
 
 import com.dodream.job.domain.Job;
+import com.dodream.job.dto.response.JobAddListDto;
 import com.dodream.job.dto.response.JobListDto;
 import com.dodream.job.dto.response.JobResponseDto;
 import com.dodream.job.exception.JobErrorCode;
 import com.dodream.job.infrastructure.JobImageUrlGenerator;
 import com.dodream.job.repository.JobRepository;
 import com.dodream.job.repository.JobSpecification;
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -39,5 +42,12 @@ public class JobService {
         Page<Job> jobs = jobRepository.findAll(spec, pageable);
 
         return jobs.map(job -> JobListDto.from(job, imageUrlGenerator));
+    }
+
+    public List<JobAddListDto> getAddJobList(){
+        return jobRepository.findAll()
+            .stream()
+            .map(job -> JobAddListDto.from(job,imageUrlGenerator))
+            .toList();
     }
 }
